Introduce `shared` and `identifier` modules
