+++ packages/graph/hash_graph/lib/graph/src/api/rest/property_type.rs
@@ -22,12 +22,12 @@ use crate::{
         patch_id_and_parse, AccountId, PersistedOntologyIdentifier, PersistedOntologyMetadata,
         PersistedPropertyType,
     },
+    shared::identifier::GraphElementIdentifier,
     store::{
         query::Expression, BaseUriAlreadyExists, BaseUriDoesNotExist, PropertyTypeStore, StorePool,
     },
     subgraph::{
-        EdgeKind, Edges, GraphElementIdentifier, GraphResolveDepths, OutwardEdge, StructuralQuery,
-        Subgraph, Vertex,
+        EdgeKind, Edges, GraphResolveDepths, OutwardEdge, StructuralQuery, Subgraph, Vertex,
     },
 };
